+++ app/src/main/java/com/example/ogrdapp/model/User.java
@@ -8,9 +8,9 @@ public class User {
     private String username;
     private long hoursOverall;
     private long  hoursToSettle;
-    private int paycheck;
+    private double paycheck;
 
-    public User(String email, String foreign_key, String surName, String userId, String userName,long hoursOverall,long hoursToSettle, int paycheck) {
+    public User(String email, String foreign_key, String surName, String userId, String userName,long hoursOverall,long hoursToSettle, double paycheck) {
         this.email = email;
         this.foreign_key = foreign_key;
         this.surName = surName;
@@ -80,11 +80,11 @@ public class User {
         this.username = username;
     }
 
-    public int getPaycheck() {
+    public double getPaycheck() {
         return paycheck;
     }
 
-    public void setPaycheck(int paycheck) {
+    public void setPaycheck(double paycheck) {
         this.paycheck = paycheck;
     }
